requirements.txt
fastapi
hnswlib
numpy
uvicorn
psycopg2-binary